+++ dgraph/cmd/root_ee.go
@@ -13,7 +13,6 @@
 package cmd
 
 import (
-	acl "github.com/dgraph-io/dgraph/ee/acl"
 	"github.com/dgraph-io/dgraph/ee/backup"
 )
 
@@ -22,6 +21,5 @@ func init() {
 	subcommands = append(subcommands,
 		&backup.Restore,
 		&backup.LsBackup,
-		&acl.CmdAcl,
 	)
 }