+++ ee/acl/utils.go
@@ -15,12 +15,9 @@ package acl
 import (
 	"encoding/json"
 
-	"github.com/dgraph-io/dgo/v2"
 	"github.com/dgraph-io/dgo/v2/protos/api"
-	"github.com/dgraph-io/dgraph/x"
 	"github.com/golang/glog"
 	"github.com/pkg/errors"
-	"github.com/spf13/viper"
 )
 
 // GetGroupIDs returns a slice containing the group ids of all the given groups.
@@ -106,8 +103,8 @@ func UnmarshalUser(resp *api.Response, userKey string) (user *User, err error) {
 // Acl represents the permissions in the ACL system.
 // An Acl can have a predicate and permission for that predicate.
 type Acl struct {
-	Predicate string `json:"predicate"`
-	Perm      int32  `json:"perm"`
+	Predicate string `json:"dgraph.rule.predicate"`
+	Perm      int32  `json:"dgraph.rule.permission"`
 }
 
 // Group represents a group in the ACL system.
@@ -115,7 +112,7 @@ type Group struct {
 	Uid     string `json:"uid"`
 	GroupID string `json:"dgraph.xid"`
 	Users   []User `json:"~dgraph.user.group"`
-	Acls    string `json:"dgraph.group.acl"`
+	Rules   []Acl  `json:"dgraph.acl.rule"`
 }
 
 // GetUid returns the UID of the group.
@@ -156,21 +153,6 @@ func UnmarshalGroups(input []byte, groupKey string) (group []Group, err error) {
 	}
 	groups := m[groupKey]
 	return groups, nil
-}
-
-// getClientWithAdminCtx creates a client by checking the --alpha, various --tls*, and --retries
-// options, and then login using groot id and password
-func getClientWithAdminCtx(conf *viper.Viper) (*dgo.Dgraph, x.CloseFunc, error) {
-	dg, closeClient := x.GetDgraphClient(conf, false)
-	err := x.GetPassAndLogin(dg, &x.CredOpt{
-		Conf:        conf,
-		UserID:      conf.GetString(gName),
-		PasswordOpt: gPassword,
-	})
-	if err != nil {
-		return nil, nil, err
-	}
-	return dg, closeClient, nil
 }
 
 // CreateUserNQuads creates the NQuads needed to store a user with the given ID and